requirements.txt
django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
requests
python-dotenv
pillow
whitenoise
